requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==42.0.0
folium==0.15.1
